Search function
Implemented the search function! This included:
Changing the index page to include the search CTA, changing the nav to
include the link, added some styles to account for radio buttons,
adding the new nav item to the active nav item js, and adding the
search function in functions.php. Wow!